Clear the compiler warnings that could hide something
463 down to 147, and every category that can indicate a bug is now empty.
What is left is GTK and Qt deprecations, which need porting rather than
substituting, so -Wall -Wextra is worth reading again.

- immodules: drop an "underline" flag that was set and never read, spell
  an empty loop body as one, delete a static declaration with no
  definition anywhere, and cast the surrounding-text limits, which are
  already known positive, before comparing them with a string length
- x11utils: hold the X modifier masks as unsigned, which is what they are
  and what they are compared against
- panel: remove an exit-polling timeout superseded by the signal handler,
  a candidate attribute builder left over from the panel giving up
  candidate rendering, and the icon and lock that went with them
- comment out unused parameter names in the signature, the convention
  already used in 130 places here, rather than adding (void) casts
- give the GObject class/instance init functions in gtkutils the
  signatures GObject calls them with, so registering a type needs no cast
  between incompatible function types
- free an ibus engine list with g_list_free_full (), whose GDestroyNotify
  is g_object_unref's own type, instead of casting it to a GFunc
- add the GTypeInfo value_table the immodules left out, the language
  table's missing locale_suffix for Interlingua, the client info field the
  panel agent left uninitialised, and initialise the socket config's
  members in declaration order
- mark the utf8 encoder's switch fallthroughs as deliberate
- bound MakeNewAtom with snprintf and size its buffer for the whole range
  of int; "_server%d_%d" fitted sixteen bytes exactly, and only because
  the sequence resets at twenty

+++ modules/Filter/scim_hanconv_filter.cpp
@@ -572,7 +572,7 @@ HanConvFilterInstance::filter_update_lookup_table (const LookupTable &table)
     } else {
         CommonLookupTable ntable;
         std::vector<WideString> labels;
-        size_t i;
+        int i;
 
         // Can be paged up.
         if (table.get_current_page_start ())
@@ -586,7 +586,8 @@ HanConvFilterInstance::filter_update_lookup_table (const LookupTable &table)
             labels.push_back (__convert (table.get_candidate_label (i), dir));
         }
 
-        if (table.get_current_page_start () + table.get_current_page_size () < table.number_of_candidates ())
+        if ((uint32) (table.get_current_page_start () +
+                      table.get_current_page_size ()) < table.number_of_candidates ())
             ntable.append_candidate (0x3400);
 
         if (table.get_current_page_start ()) {
